make relationship resolvers more powerful
